Fix issues with tests
